app/requirements.txt
Flask>=1.1.2
twilio>=6.43.0
schedule>=0.6.0
python-dotenv>=0.10.3
